OCR DB config validation (#3639)
- FKs for ocr key bundles and transmitter addresses
- Human readable errors for non-existent ocr key bundles, transmitter addresses and peer ids
- Split the config tracker and transmitter into separate objects
