+++ core/services/pipeline/runner_test.go
@@ -334,7 +334,7 @@ ds1 -> ds1_parse;
 			Pipeline: *pipeline.NewTaskDAG(),
 		}
 		s := fmt.Sprintf(minimalNonBootstrapTemplate, cltest.NewEIP55Address(), ek.PeerID, cltest.DefaultKey, "chain.link:101", kb.ID)
-		_, err = services.ValidatedOracleSpec(s)
+		_, err = services.ValidatedOracleSpecToml(s)
 		require.NoError(t, err)
 		err = toml.Unmarshal([]byte(s), &os)
 		require.NoError(t, err)
@@ -385,7 +385,7 @@ ds1 -> ds1_parse;
 			Pipeline: *pipeline.NewTaskDAG(),
 		}
 		s := fmt.Sprintf(minimalBootstrapTemplate, cltest.NewEIP55Address(), ek.PeerID)
-		_, err = services.ValidatedOracleSpec(s)
+		_, err = services.ValidatedOracleSpecToml(s)
 		require.NoError(t, err)
 		err = toml.Unmarshal([]byte(s), &os)
 		require.NoError(t, err)
