+++ core/web/ocr_job_specs_controller.go
@@ -7,6 +7,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/chainlink/core/services"
 	"github.com/smartcontractkit/chainlink/core/services/chainlink"
+	"github.com/smartcontractkit/chainlink/core/services/job"
 	"github.com/smartcontractkit/chainlink/core/services/offchainreporting"
 	"github.com/smartcontractkit/chainlink/core/store/models"
 	"github.com/smartcontractkit/chainlink/core/store/orm"
@@ -64,7 +65,7 @@ func (ocrjsc *OCRJobSpecsController) Create(c *gin.Context) {
 		jsonAPIError(c, http.StatusUnprocessableEntity, err)
 		return
 	}
-	jobSpec, err := services.ValidatedOracleSpec(request.TOML)
+	jobSpec, err := services.ValidatedOracleSpecToml(request.TOML)
 	if err != nil {
 		jsonAPIError(c, http.StatusBadRequest, err)
 		return
@@ -77,6 +78,10 @@ func (ocrjsc *OCRJobSpecsController) Create(c *gin.Context) {
 
 	jobID, err := ocrjsc.App.AddJobV2(c.Request.Context(), jobSpec)
 	if err != nil {
+		if errors.Cause(err) == job.ErrNoSuchKeyBundle || errors.Cause(err) == job.ErrNoSuchPeerID || errors.Cause(err) == job.ErrNoSuchTransmitterAddress {
+			jsonAPIError(c, http.StatusBadRequest, err)
+			return
+		}
 		jsonAPIError(c, http.StatusInternalServerError, err)
 		return
 	}
