+++ core/web/ocr_job_specs_controller_test.go
@@ -9,6 +9,8 @@ import (
 	"net/http"
 	"testing"
 
+	"github.com/smartcontractkit/chainlink/core/services/job"
+
 	"github.com/BurntSushi/toml"
 	"github.com/smartcontractkit/chainlink/core/internal/cltest"
 	"github.com/smartcontractkit/chainlink/core/services/offchainreporting"
@@ -19,19 +21,56 @@ import (
 )
 
 func TestOCRJobSpecsController_Create_ValidationFailure(t *testing.T) {
-	_, client, cleanup := setupOCRJobSpecsControllerTests(t)
-	defer cleanup()
+	var (
+		contractAddress    = cltest.NewEIP55Address()
+		monitoringEndpoint = "chain.link:101"
+	)
 
-	body, _ := json.Marshal(models.CreateOCRJobSpecRequest{
-		TOML: string(cltest.MustReadFile(t, "testdata/oracle-spec-invalid-key.toml")),
-	})
-	resp, cleanup := client.Post("/v2/ocr/specs", bytes.NewReader(body))
-	defer cleanup()
-	assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
-
-	b, err := ioutil.ReadAll(resp.Body)
-	require.NoError(t, err)
-	assert.Equal(t, "{\"errors\":[{\"detail\":\"isBootstrapPeer is not defined\"}]}", string(b))
+	var tt = []struct {
+		name        string
+		pid         models.PeerID
+		kb          models.Sha256Hash
+		ta          models.EIP55Address
+		expectedErr error
+	}{
+		{
+			name:        "invalid keybundle",
+			pid:         models.PeerID(cltest.DefaultP2PPeerID),
+			kb:          models.Sha256Hash(cltest.Random32Byte()),
+			ta:          cltest.DefaultKeyAddressEIP55,
+			expectedErr: job.ErrNoSuchKeyBundle,
+		},
+		{
+			name:        "invalid peerID",
+			pid:         models.PeerID(cltest.NonExistentP2PPeerID),
+			kb:          cltest.DefaultOCRKeyBundleIDSha256,
+			ta:          cltest.DefaultKeyAddressEIP55,
+			expectedErr: job.ErrNoSuchPeerID,
+		},
+		{
+			name:        "invalid transmitter address",
+			pid:         models.PeerID(cltest.DefaultP2PPeerID),
+			kb:          cltest.DefaultOCRKeyBundleIDSha256,
+			ta:          cltest.NewEIP55Address(),
+			expectedErr: job.ErrNoSuchTransmitterAddress,
+		},
+	}
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			_, client, cleanup := setupOCRJobSpecsControllerTests(t)
+			defer cleanup()
+			sp := cltest.MinimalOCRNonBootstrapSpec(contractAddress, tc.ta, tc.pid, monitoringEndpoint, tc.kb)
+			body, _ := json.Marshal(models.CreateOCRJobSpecRequest{
+				TOML: sp,
+			})
+			resp, cleanup := client.Post("/v2/ocr/specs", bytes.NewReader(body))
+			defer cleanup()
+			assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
+			b, err := ioutil.ReadAll(resp.Body)
+			require.NoError(t, err)
+			assert.Contains(t, string(b), tc.expectedErr.Error())
+		})
+	}
 }
 
 func TestOCRJobSpecsController_Create_HappyPath(t *testing.T) {
